// src/core/types/image.type.ts
export interface IPImage {
  id: "126d854f-794e-49cc-baf1-a5c56f6cd4aa";
  publicId: "portfolio/ammpbabvvu20fwkcg83r";
  url: "http://res.cloudinary.com/dijghadni/image/upload/v1764225338/portfolio/ammpbabvvu20fwkcg83r.png";
  secureUrl: "https://res.cloudinary.com/dijghadni/image/upload/v1764225338/portfolio/ammpbabvvu20fwkcg83r.png";
  folder: "portfolio";
  name: "file";
  originalFilename: "file";
  format: "png";
  width: 215;
  height: 215;
  bytes: 7587;
  resourceType: "image";
  uploadedAt: "2025-11-27T06:35:38.000Z";
  provider: "cloudinary";
  tags: [];
  createdAt: "2025-11-27T06:35:41.140Z";
}